# docs/knowledge/backend/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/users/staff/by-branch/{branchId}:
    get:
      operationId: UsersController_findStaffByBranch
      parameters:
        - name: branchId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Get all staff members for a specific branch
      tags: &a6
        - Users
  /api/users/departments/by-branch/{branchId}:
    get:
      operationId: UsersController_getDepartmentsByBranch
      parameters:
        - name: branchId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Get all unique departments for a specific branch
      tags: *a6
  /api/users/staff/by-department/{branchId}/{department}:
    get:
      operationId: UsersController_findStaffByDepartment
      parameters:
        - name: branchId
          required: true
          in: path
          schema:
            type: string
        - name: department
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Get staff members for a specific department in a branch
      tags: *a6
  /api/users/staff/search:
    get:
      operationId: UsersController_searchStaff
      parameters:
        - name: branchId
          required: true
          in: query
          schema:
            type: string
        - name: query
          required: true
          in: query
          schema:
            type: string
        - name: department
          required: true
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      summary: Search staff members by name in a specific branch
      tags: *a6
  /api/users:
    post:
      description: "- SUPER_ADMIN can create any user.\\n- CHAIN_ADMIN can create
        BRANCH_ADMIN, STAFF, and SECURITY in their chain.\\n- BRANCH_ADMIN can
        create STAFF and SECURITY in their branch."
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
      security:
        - access-token: []
      summary: Create a new user (any role based on permissions)
      tags: *a6
    get:
      operationId: UsersController_findAll
      parameters:
        - name: role
          required: false
          in: query
          description: Filter users by a specific role (e.g., CHAIN_ADMIN).
          schema:
            type: string
            enum:
              - SUPER_ADMIN
              - CHAIN_ADMIN
              - BRANCH_ADMIN
              - SECURITY_SUPERVISOR
              - SECURITY
              - STAFF
        - name: isActive
          required: false
          in: query
          description: Filter users by their active status (true or false).
          schema:
            default: true
            type: boolean
        - name: branchId
          required: false
          in: query
          description: Filter users by a specific branch ID.
          schema:
            type: string
        - name: chainId
          required: false
          in: query
          description: Filter users by a specific hospital chain ID.
          schema:
            type: string
        - name: department
          required: false
          in: query
          description: Filter STAFF users by a specific department.
          schema:
            type: string
            enum:
              - GENERAL_MEDICINE
              - CARDIOLOGY
              - NEUROLOGY
              - PULMONOLOGY
              - NEPHROLOGY
              - GASTROENTEROLOGY
              - ENDOCRINOLOGY
              - ONCOLOGY
              - HEMATOLOGY
              - DERMATOLOGY
              - INFECTIOUS_DISEASE
              - RHEUMATOLOGY
              - OPHTHALMOLOGY
              - DENTISTRY
              - PAIN_MANAGEMENT
              - GENERAL_SURGERY
              - ORTHOPEDICS
              - NEUROSURGERY
              - CARDIOTHORACIC_SURGERY
              - ENT_SURGERY
              - PLASTIC_RECONSTRUCTIVE_SURGERY
              - UROLOGY
              - OBSTETRICS_GYNECOLOGY
              - PEDIATRICS
              - NEONATOLOGY
              - EMERGENCY
              - INTENSIVE_CARE_UNIT
              - CRITICAL_CARE_UNIT
              - TRAUMA_CENTER
              - RADIOLOGY
              - PATHOLOGY
              - PHARMACY
              - BLOOD_BANK
              - ANESTHESIOLOGY
              - RESPIRATORY_CARE
              - CENTRAL_SUPPLY
              - NUTRITION_AND_DIETETICS
              - MEDICAL_RECORDS
              - SECURITY
              - ADMINISTRATION
              - HUMAN_RESOURCES
              - FINANCE
              - IT_SUPPORT
              - LEGAL_COMPLIANCE
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: List and filter all users based on permissions
      tags: *a6
  /api/users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Get a specific user by ID
      tags: *a6
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Update a user by ID
      tags: *a6
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security:
        - access-token: []
      summary: Delete a user by ID
      tags: *a6
...
info:
  title: Hospital Visitor Tracking System
  description: API documentation for the Hospital Visitor Tracking System
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - CHAIN_ADMIN
            - BRANCH_ADMIN
            - SECURITY_SUPERVISOR
            - SECURITY
            - STAFF
          description: The role of the user to be created.
        userType:
          type: string
          enum:
            - DOCTOR
            - NURSE
            - SURGEON
            - PHYSICIAN_ASSISTANT
            - RESIDENT
            - INTERN
            - CONSULTANT
            - CLINICAL_RESEARCHER
            - SPECIALIST
            - ANESTHESIOLOGIST
            - RADIOLOGIST
            - PATHOLOGIST
            - PSYCHIATRIST
            - PHYSIOTHERAPIST
            - HEAD_NURSE
            - STAFF_NURSE
            - NURSE_PRACTITIONER
            - MIDWIFE
            - NURSING_ASSISTANT
            - EMERGENCY_PHYSICIAN
            - PARAMEDIC
            - EMT
            - ICU_SPECIALIST
            - TRAUMA_SURGEON
            - PHARMACIST
            - LAB_TECHNICIAN
            - LAB_ASSISTANT
            - DIETICIAN
            - NUTRITIONIST
            - OCCUPATIONAL_THERAPIST
            - SPEECH_THERAPIST
            - RESPIRATORY_THERAPIST
            - HOSPITAL_ADMINISTRATOR
            - DEPARTMENT_HEAD
            - MEDICAL_RECORDS_STAFF
            - BILLING_STAFF
            - RECEPTIONIST
            - RADIOLOGY_TECHNICIAN
            - ECG_TECHNICIAN
            - MEDICAL_TECHNOLOGIST
            - BIOMEDICAL_ENGINEER
            - SECURITY_CHIEF
            - SECURITY_GUARD
            - FACILITY_MANAGER
            - HOUSEKEEPING_STAFF
            - CLEANING_STAFF
            - MAINTENANCE_STAFF
            - IT_SUPPORT
            - PHARMACY_TECHNICIAN
            - VOLUNTEER_COORDINATOR
            - PATIENT_COORDINATOR
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        location:
          type: string
        hospitalChainId:
          type: string
          description: Required for all roles except SUPER_ADMIN.
        branchId:
          type: string
          description: Required for all roles that belong to a branch (e.g., STAFF,
            BRANCH_ADMIN).
      required:
        - name
        - phone
        - role
        - hospitalChainId
    UpdateUserDto:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - CHAIN_ADMIN
            - BRANCH_ADMIN
            - SECURITY_SUPERVISOR
            - SECURITY
            - STAFF
        userType:
          type: string
          enum:
            - DOCTOR
            - NURSE
            - SURGEON
            - PHYSICIAN_ASSISTANT
            - RESIDENT
            - INTERN
            - CONSULTANT
            - CLINICAL_RESEARCHER
            - SPECIALIST
            - ANESTHESIOLOGIST
            - RADIOLOGIST
            - PATHOLOGIST
            - PSYCHIATRIST
            - PHYSIOTHERAPIST
            - HEAD_NURSE
            - STAFF_NURSE
            - NURSE_PRACTITIONER
            - MIDWIFE
            - NURSING_ASSISTANT
            - EMERGENCY_PHYSICIAN
            - PARAMEDIC
            - EMT
            - ICU_SPECIALIST
            - TRAUMA_SURGEON
            - PHARMACIST
            - LAB_TECHNICIAN
            - LAB_ASSISTANT
            - DIETICIAN
            - NUTRITIONIST
            - OCCUPATIONAL_THERAPIST
            - SPEECH_THERAPIST
            - RESPIRATORY_THERAPIST
            - HOSPITAL_ADMINISTRATOR
            - DEPARTMENT_HEAD
            - MEDICAL_RECORDS_STAFF
            - BILLING_STAFF
            - RECEPTIONIST
            - RADIOLOGY_TECHNICIAN
            - ECG_TECHNICIAN
            - MEDICAL_TECHNOLOGIST
            - BIOMEDICAL_ENGINEER
            - SECURITY_CHIEF
            - SECURITY_GUARD
            - FACILITY_MANAGER
            - HOUSEKEEPING_STAFF
            - CLEANING_STAFF
            - MAINTENANCE_STAFF
            - IT_SUPPORT
            - PHARMACY_TECHNICIAN
            - VOLUNTEER_COORDINATOR
            - PATIENT_COORDINATOR
        department:
          type: string
          enum:
            - GENERAL_MEDICINE
            - CARDIOLOGY
            - NEUROLOGY
            - PULMONOLOGY
            - NEPHROLOGY
            - GASTROENTEROLOGY
            - ENDOCRINOLOGY
            - ONCOLOGY
            - HEMATOLOGY
            - DERMATOLOGY
            - INFECTIOUS_DISEASE
            - RHEUMATOLOGY
            - OPHTHALMOLOGY
            - DENTISTRY
            - PAIN_MANAGEMENT
            - GENERAL_SURGERY
            - ORTHOPEDICS
            - NEUROSURGERY
            - CARDIOTHORACIC_SURGERY
            - ENT_SURGERY
            - PLASTIC_RECONSTRUCTIVE_SURGERY
            - UROLOGY
            - OBSTETRICS_GYNECOLOGY
            - PEDIATRICS
            - NEONATOLOGY
            - EMERGENCY
            - INTENSIVE_CARE_UNIT
            - CRITICAL_CARE_UNIT
            - TRAUMA_CENTER
            - RADIOLOGY
            - PATHOLOGY
            - PHARMACY
            - BLOOD_BANK
            - ANESTHESIOLOGY
            - RESPIRATORY_CARE
            - CENTRAL_SUPPLY
            - NUTRITION_AND_DIETETICS
            - MEDICAL_RECORDS
            - SECURITY
            - ADMINISTRATION
            - HUMAN_RESOURCES
            - FINANCE
            - IT_SUPPORT
            - LEGAL_COMPLIANCE
        location:
          type: string
        hospitalChainId:
          type: string
        branchId:
          type: string
        isActive:
          type: boolean
...